Make READ THE INTRODUCTION button on homepage clickable
The "call to action" button at the bottom of the homepage is currently unclickable, since it has a negative top margin and is thus "hidden" behind the bottom margin of the terminal div above it. I opted to remove the negative top margin and instead diminish the bottom margin of the terminal div. This approach ends up adding a bit more padding between the "DNS Query Interface" h2 and the marketing div above it, but I think overall the look is in-line with the aesthetics of the page.

I'm fairly unfamiliar with HTML / CSS, so if there is a better way to do what I've done, I'd love to hear it. I've tested this in Chrome and Safari.